refactor: move WindowUtil to root namespace
- Relocate WindowUtil class from Eduard.Cryptography to Eduard namespace
- Improve namespace organization for better code accessibility
- Enable broader utilization of windowing utilities across the library
