+++ Eduard/Cryptography/Polynomial.cs
@@ -2,6 +2,7 @@
 using System.Diagnostics;
 using System.Collections.Generic;
 using System.Security.Cryptography;
+using Eduard;
 
 namespace Eduard.Cryptography
 {
